requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
biomass==0.14.0
blockbuster==1.5.25
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.1.1
exa-py==1.14.17
filetype==1.2.0
fonttools==4.59.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.104.0
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.26.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
groq==0.30.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-community==0.3.27
langchain-core==0.3.69
langchain-deepseek==0.1.3
langchain-google-genai==2.1.8
langchain-google-vertexai==2.0.27
langchain-groq==0.3.6
langchain-openai==0.3.28
langchain-tavily==0.2.10
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-api==0.2.98
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.5
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.6.0
langgraph-sdk==0.1.73
langsmith==0.4.8
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
markdownify==1.1.0
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numexpr==2.11.0
numpy==2.2.6
openai==1.97.0
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
scipy==1.16.0
seaborn==0.13.2
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.47.2
structlog==25.4.0
tabulate==0.9.0
tavily-python==0.7.10
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
